Create Smart Saline Bottle System
This is a code for smart saline bottle system made by our team in Engineering exploration class. It detects the amount of electrolyte in the saline bottle and messages nurses and doctors about the critical level of the electrolyte in saline bottle.
Note: You need to join all components connected in order to run this code.